requirements/dev.txt
-r base.txt

django-debug-toolbar==1.5
https://github.com/omarish/django-cprofile-middleware/tarball/0.2

# Mac/OSX only
bpython==0.15.0

# for testing
pytest-django==2.9.1
